Reimplemented poll() syscall handler.
The problem with the poll() syscall is that the gpiointr driver does not
maintain any data buffers which could be checked for emptiness in the
handler. Instead an interrupt is a single events, that is already gone
after a poll() or select() syscall returns.

This is a quick-and-dirty implementation. For the long term an extra
thread using the observer pattern should be implemented to handle these
syscalls appropriately.