Fixed locking during interrupt configuration.
The interrupt configuration cannot be properly protected by the mutex of the
interrupt configuration structure, because mutexes are not allowed to be
held during calls to several necessary bus functions. This issue is evaded
by a additional locking variable that itself is locked by the mutex of the
interrupt configuration struct. When the mutex is acquired somewhere outside
of the interrupt allocate and release functions the value of the variable is
checked first, before taking any other actions.